# docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.AdminCreateUserRequest:
    properties:
      email:
        type: string
      password:
        maxLength: 128
        minLength: 8
        type: string
      role:
        enum:
        - user
        - admin
        type: string
      username:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - email
    - password
    - username
    type: object
...
  dto.UserResponse:
    properties:
      email:
        type: string
      role:
        type: string
      user_id:
        type: integer
      username:
        type: string
    type: object
  response.ErrorResponse:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: Bad request
        type: string
      request_id:
        example: abc-123
        type: string
    type: object
  response.PaginatedResponse:
    properties:
      code:
        example: 200
        type: integer
      data: {}
      message:
        example: Success
        type: string
      pagination:
        $ref: '#/definitions/response.Pagination'
      request_id:
        example: abc-123
        type: string
    type: object
  response.Pagination:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  response.SuccessResponse:
    properties:
      code:
        example: 200
        type: integer
      data: {}
      message:
        example: Success
        type: string
      request_id:
        example: abc-123
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A high-performance todo list microservice built with Go and Clean Architecture
    principles.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Todo List API
  version: "1.0"
paths:
  /admin/todos:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of all todos with optional filters (admin
        only)
      parameters:
      - default: 1
        description: Page number
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - description: Filter by status
        enum:
        - not_started
        - in_progress
        - completed
        in: query
        name: status
        type: string
      - description: Filter by priority
        enum:
        - low
        - medium
        - high
        in: query
        name: priority
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Todos retrieved successfully
          schema:
            $ref: '#/definitions/response.PaginatedResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Forbidden - admin role required
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: List all todos (Admin)
      tags:
      - Admin
  /admin/todos/{id}:
    delete:
      consumes:
      - application/json
      description: Delete any todo by ID, regardless of ownership (admin only)
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Todo deleted successfully
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Invalid todo ID
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Forbidden - admin role required
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Todo not found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Delete any todo (Admin)
      tags:
      - Admin
  /admin/users:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of all users (admin only)
      parameters:
      - default: 1
        description: Page number
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Users retrieved successfully
          schema:
            $ref: '#/definitions/response.PaginatedResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Forbidden - admin role required
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: List all users (Admin)
      tags:
      - Admin
    post:
      consumes:
      - application/json
      description: Create a new user account (admin only)
      parameters:
      - description: User creation details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AdminCreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User created successfully
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "400":
          description: Invalid request format or validation error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Forbidden - admin role required
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "409":
          description: Username or email already exists
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Create a user (Admin)
      tags:
      - Admin
  /admin/users/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a user by ID (admin only)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User deleted successfully
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Invalid user ID
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Forbidden - admin role required
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Delete user (Admin)
      tags:
      - Admin
    get:
      consumes:
      - application/json
      description: Retrieve a specific user by ID (admin only)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User retrieved successfully
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "400":
          description: Invalid user ID
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Forbidden - admin role required
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Get user by ID (Admin)
      tags:
      - Admin
...
securityDefinitions:
  Bearer:
    description: Enter the token with the `Bearer ` prefix, e.g. "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"